refactor(Table): editable, creation and bulk editing are moved to separate hooks
