+++ packages/common-frontend/src/components/Descriptions/descriptionTypes.ts
@@ -27,11 +27,14 @@ export type TDescriptionsProps<Entity, CreateDto, UpdateDto, TPathParams = objec
   mainTitle?: ProColumns<Entity>['title'] | null,
   entity?: Partial<Entity>,
   getOne?: ({}: TGetOneParams & TPathParams) => Promise<Entity | null>,
-  onUpdate?: ({}: Record<keyof Entity, string> & { requestBody: UpdateDto } & TPathParams) => Promise<Entity>,
-  onDelete?: ({}: Record<keyof Entity, string> & TPathParams) => Promise<void>,
+  onUpdate?: ({}: Partial<Entity> & {
+    requestBody: UpdateDto,
+    index?: number,
+  } & TPathParams) => Promise<Entity>,
+  onDelete?: ({}: Partial<Entity> & TPathParams) => Promise<void>,
   pathParams?: TPathParams,
   idColumnName?: string & keyof Entity,
-  entityToUpdateDto?: (entity: Partial<Entity>) => UpdateDto,
+  entityToUpdateDto?: (entity: Entity) => UpdateDto,
   createNewDefaultParams?: Partial<Entity>,
   afterSave?: (record: Entity) => Promise<void>,
   actionRef?: MutableRefObject<ActionType | undefined>,
