+++ packages/common-frontend/src/components/Table/index.ts
@@ -4,3 +4,4 @@ export * from './tableTools';
 export * from './tableTypes';
 export * from './useColumnsSets';
 export { default as useColumnsSets } from './useColumnsSets';
+export * from "./useCreation";
